apps/api/requirements.txt
# FastAPI & Web Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# AI Providers
openai==1.12.0
anthropic==0.18.1
google-cloud-aiplatform==1.43.0
google-auth==2.27.0
tiktoken==0.6.0

# AWS
boto3==1.34.41
botocore==1.34.41

# Document Generation
python-docx==1.1.0
python-pptx==0.6.23
jinja2==3.1.3
playwright==1.41.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx-mock==0.13.0

# Code Quality
ruff==0.2.1
black==24.1.1
mypy==1.8.0

# Utilities
python-json-logger==2.0.7

